Split `tasks/utils.py` into v1 and v2
