+++ fractal_server/tasks/v1/background_operations.py
@@ -6,13 +6,13 @@ import json
 from pathlib import Path
 from shutil import rmtree as shell_rmtree
 
-from ..utils import _init_venv_v1
 from ..utils import _normalize_package_name
 from ..utils import get_collection_log
 from ..utils import get_collection_path
 from ..utils import get_log_path
 from ..utils import slugify_task_name
 from ._TaskCollectPip import _TaskCollectPip
+from .utils import _init_venv_v1
 from fractal_server.app.db import DBSyncSession
 from fractal_server.app.db import get_sync_db
 from fractal_server.app.models.v1 import State
